transferred event listeners to main.js, made separate task comp create funcs for edit feature, created chatFunctions file, removed chat.js, added chat imports to main.js, called showelement func for chatWrapper
