+++ src/index.html
@@ -13,7 +13,6 @@
   <article id="container"></article>
 
   <script src="./scripts/main.js" type="module"></script>
-  <script type="module" src="scripts/chat.js"></script>
 </body>
 
 </html>
